Implement per package targets at 1 level
