+++ tools/aapt_lite/BUILD
@@ -1,34 +1,13 @@
 load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
 load("@grab_bazel_common//tools/test:test.bzl", "grab_kt_jvm_test")
 
-kt_jvm_library(
-    name = "aapt_lite_lib",
-    srcs = glob([
-        "src/main/java/**/*.kt",
-    ]),
-    visibility = [
-        "//visibility:private",
-    ],
-    deps = [
-        "//:dagger",
-        "//tools/worker:worker_lib",
-        "@bazel_common_maven//:com_github_ajalt_clikt",
-        "@bazel_common_maven//:com_squareup_javapoet",
-        "@bazel_common_maven//:com_squareup_moshi_moshi",
-        "@bazel_common_maven//:net_sf_kxml_kxml2",
-        "@bazel_common_maven//:xmlpull_xmlpull",
-        "@com_github_jetbrains_kotlin//:kotlin-stdlib-jdk8",
-        "@maven//:androidx_databinding_databinding_common",
-    ],
-)
-
 grab_kt_jvm_test(
     name = "aapt_lite_test",
     srcs = glob([
         "src/test/java/**/*.kt",
     ]),
     deps = [
-        ":aapt_lite_lib",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt",
         "@bazel_common_maven//:junit_junit",
         "@com_github_jetbrains_kotlin//:kotlin-test",
     ],
@@ -41,6 +20,6 @@ java_binary(
         "//visibility:public",
     ],
     runtime_deps = [
-        ":aapt_lite_lib",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt",
     ],
 )
