+++ tools/aapt_lite/src/main/java/com/grab/aapt/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "aapt",
+    srcs = glob([
+        "*.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/brclass",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di:aapt_scope",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/mapper",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/rclass",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/util/jars",
+        "//tools/worker:worker_lib",
+    ],
+)
