+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "binding",
+    srcs = glob([
+        "**/*.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding/model",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/common",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di:aapt_scope",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/util",
+        "@bazel_common_maven//:com_squareup_javapoet",
+        "@bazel_common_maven//:com_squareup_moshi_moshi",
+    ],
+)
