+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding/model/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "model",
+    srcs = glob([
+        "Models.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/util",
+        "@bazel_common_maven//:com_squareup_javapoet",
+    ],
+)
