+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding/store/LayoutTypeStore.kt
@@ -16,7 +16,7 @@
 
 package com.grab.aapt.databinding.binding.store
 
-import com.grab.aapt.AaptScope
+import com.grab.aapt.databinding.di.AaptScope
 import com.grab.aapt.databinding.common.CLASS_INFOS
 import com.grab.aapt.databinding.common.LAYOUT_FILES
 import com.grab.aapt.databinding.common.PACKAGE_NAME
