+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/brclass/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "brclass",
+    srcs = glob([
+        "BrClassGenerator.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding/model",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/common",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di:aapt_scope",
+        "@bazel_common_maven//:com_squareup_javapoet",
+    ],
+)
