+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/common/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "common",
+    srcs = glob([
+        "*.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+)
