+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di/AaptScope.kt
@@ -1,0 +1,6 @@
+package com.grab.aapt.databinding.di
+
+import javax.inject.Scope
+
+@Scope
+annotation class AaptScope
