+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "aapt_scope",
+    srcs = glob([
+        "AaptScope.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+    ],
+)
