+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/mapper/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "mapper",
+    srcs = glob([
+        "*.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/binding",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di:aapt_scope",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/util/jars",
+        "@bazel_common_maven//:com_github_ajalt_clikt",
+        "@com_github_jetbrains_kotlin//:kotlin-stdlib-jdk8",
+    ],
+)
