+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/rclass/parser/XmlEntry.kt
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package com.grab.aapt.databinding.util
-
-import com.grab.aapt.databinding.rclass.parser.XmlTypeValues
+package com.grab.aapt.databinding.rclass.parser
 
 interface XmlEntry {
     /**
