+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/rclass/parser/xml/ArrayParser.kt
@@ -20,7 +20,7 @@ import com.grab.aapt.databinding.rclass.parser.ParserResult
 import com.grab.aapt.databinding.rclass.parser.RFieldEntry
 import com.grab.aapt.databinding.rclass.parser.ResourceFileParser
 import com.grab.aapt.databinding.rclass.parser.Type
-import com.grab.aapt.databinding.util.XmlEntry
+import com.grab.aapt.databinding.rclass.parser.XmlEntry
 import javax.inject.Inject
 
 /**
