+++ tools/aapt_lite/src/main/java/com/grab/aapt/databinding/util/jars/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "jars",
+    srcs = glob([
+        "*.kt",
+    ]),
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "//:dagger",
+        "//tools/aapt_lite/src/main/java/com/grab/aapt/databinding/di:aapt_scope",
+    ],
+)
