+++ tools/aapt_lite/src/test/java/com/grab/aapt/databinding/rclass/parser/DefaultRTxtParserTest.kt
@@ -16,9 +16,9 @@
 
 package com.grab.aapt.databinding.rclass.parser
 
-import com.grab.aapt.databinding.binding.parser.Binding
-import com.grab.aapt.databinding.binding.parser.BindingType
-import com.grab.aapt.databinding.binding.parser.LayoutBindingData
+import com.grab.aapt.databinding.binding.model.Binding
+import com.grab.aapt.databinding.binding.model.BindingType
+import com.grab.aapt.databinding.binding.model.LayoutBindingData
 import com.grab.aapt.databinding.common.BaseBindingStubTest
 import com.grab.aapt.databinding.rclass.parser.Type.*
 import com.squareup.javapoet.ClassName
