+++ tools/binding-adapter-bridge/BUILD.bazel
@@ -12,7 +12,6 @@ kt_jvm_library(
         "@bazel_common_maven//:com_google_auto_service_auto_service",
         "@bazel_common_maven//:com_google_guava_guava",
         "@bazel_common_maven//:com_squareup_javapoet",
-        "@com_github_jetbrains_kotlin//:kotlin-stdlib-jdk8",
         "@maven//:androidx_databinding_databinding_adapters",
     ],
 )
